refactor: Move admin-related files to app
