--- src/app/redirect/page.tsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const redirect = () => {
-  return (
-    <div>
-      <h1>Fuiste redirigido xd</h1>
-    </div>
-  )
-}
-
-export default redirect
