--- src/app/utils/roles.ts
@@ -1,7 +1,0 @@
-import { Roles } from '../types/globals'
-import { auth } from '@clerk/nextjs/server'
-
-export const checkRole = async (role: Roles) => {
-    const { sessionClaims } = await auth()
-    return sessionClaims?.metadata.role === role
-}
