+++ src/types/globals.d.ts
@@ -1,7 +1,7 @@
 export { }
 
 // Create a type for the roles
-export type Roles = 'admin' | 'moderator'
+export type Roles = 'admin' | 'moderator' | 'user' | null
 
 declare global {
     interface CustomJwtSessionClaims {
